# DTO/AccountDTO.cs
namespace CameraBase.DTO
{
    public class AccountDTO
    {

        public string Password { get; set; }

        public string AccountEmail { get; set; }

        public string Phone { get; set; }

        public string fullName { get; set; }

        public string Image { get; set; }

        public int RoleID  { get; set; }
    }
}