# bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "CameraBased API",
    "version": "v1"
  },
  "paths": {
...
    "/CarLocator/GetById/{id}": {
      "get": {
        "tags": [
          "CarLocator"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Carlocator"
            }
          }
        }
      }
    },
    "/CarLocator/Create": {
      "post": {
        "tags": [
          "CarLocator"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarLocatorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Carlocator"
            }
          }
        }
      }
    },
    "/CarLocator/Update/{id}": {
      "put": {
        "tags": [
          "CarLocator"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarLocatorDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/CarLocator/Delete/{id}": {
      "delete": {
        "tags": [
          "CarLocator"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/CarManagement/Get": {
      "get": {
        "tags": [
          "CarManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CarManagement"
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/CarManagement/GetByName/{name}": {
      "get": {
        "tags": [
          "CarManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CarManagement"
              }
            }
          }
        }
      }
    },
    "/CarManagement/GetById/{id}": {
      "get": {
        "tags": [
          "CarManagement"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/CarManagement"
              }
            }
          }
        }
      }
    },
    "/CarManagement/ChangeStatus/{id}": {
      "get": {
        "tags": [
          "CarManagement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/CarManagement/Create": {
      "post": {
        "tags": [
          "CarManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarManagementDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/CarManagement"
            }
          }
        }
      }
    },
    "/CarManagement/Update/{id}": {
      "put": {
        "tags": [
          "CarManagement"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CarManagementDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/CarManagement/Delete/{id}": {
      "delete": {
        "tags": [
          "CarManagement"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
    "/Statistical/TotalOfAccount": {
      "get": {
        "tags": [
          "Statistical"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Statistical/TotalOfSubAccount": {
      "get": {
        "tags": [
          "Statistical"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Statistical/TotalOfHistory": {
      "get": {
        "tags": [
          "Statistical"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
    "/Statistical/TotalOfCar": {
      "get": {
        "tags": [
          "Statistical"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "security": [
          {
            "oauth2": [ ]
          }
        ]
      }
    },
...
  },
  "definitions": {
    "Account": {
      "required": [
        "accounName",
        "password"
      ],
      "type": "object",
      "properties": {
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "accounName": {
          "maxLength": 100,
          "type": "string"
        },
        "password": {
          "maxLength": 100,
          "type": "string"
        },
        "accountEmail": {
          "maxLength": 200,
          "type": "string"
        },
        "phone": {
          "maxLength": 20,
          "type": "string"
        },
        "fullName": {
          "maxLength": 100,
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "roleID": {
          "format": "int32",
          "type": "integer"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "notifiHistories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotifiHistory"
          }
        },
        "subAccounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/SubAccount"
          }
        },
        "refreshToken": {
          "type": "string"
        },
        "tokenCreated": {
          "format": "date-time",
          "type": "string"
        },
        "tokenExpires": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Carlocator": {
      "type": "object",
      "properties": {
        "carLocatorID": {
          "format": "int32",
          "type": "integer"
        },
        "location": {
          "type": "string"
        },
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carManagement": {
          "$ref": "#/definitions/CarManagement"
        }
      },
      "additionalProperties": false
    },
    "CarLocatorDTO": {
      "type": "object",
      "properties": {
        "location": {
          "type": "string"
        },
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "CarManagement": {
      "type": "object",
      "properties": {
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carName": {
          "maxLength": 100,
          "type": "string"
        },
        "carColor": {
          "maxLength": 100,
          "type": "string"
        },
        "licensePlates": {
          "maxLength": 100,
          "type": "string"
        },
        "carBrand": {
          "maxLength": 100,
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "notifiHistories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NotifiHistory"
          }
        },
        "carlocators": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Carlocator"
          }
        }
      },
      "additionalProperties": false
    },
    "CarManagementDTO": {
      "type": "object",
      "properties": {
        "carName": {
          "type": "string"
        },
        "carColor": {
          "type": "string"
        },
        "licensePlates": {
          "type": "string"
        },
        "carBrand": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "NotifiHistory": {
      "type": "object",
      "properties": {
        "historyID": {
          "format": "int32",
          "type": "integer"
        },
        "historyName": {
          "maxLength": 100,
          "type": "string"
        },
        "historyDate": {
          "format": "date-time",
          "type": "string"
        },
        "historyStatus": {
          "type": "boolean"
        },
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "account": {
          "$ref": "#/definitions/Account"
        },
        "carManagementID": {
          "format": "int32",
          "type": "integer"
        },
        "carManagement": {
          "$ref": "#/definitions/CarManagement"
        }
      },
      "additionalProperties": false
    },
    "Role": {
      "required": [
        "roleName"
      ],
      "type": "object",
      "properties": {
        "roleID": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "maxLength": 30,
          "type": "string"
        },
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Account"
          }
        }
      },
      "additionalProperties": false
    },
    "SubAccount": {
      "type": "object",
      "properties": {
        "subAccountID": {
          "format": "int32",
          "type": "integer"
        },
        "subAccountName": {
          "type": "string"
        },
        "subAccountPhone": {
          "type": "string"
        },
        "accountID": {
          "format": "int32",
          "type": "integer"
        },
        "account": {
          "$ref": "#/definitions/Account"
        }
      },
      "additionalProperties": false
    },
...
  },
  "securityDefinitions": {
    "oauth2": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Standard Authorization header uisng the Bearer scheme (\"bearer {token}\")"
    }
  }
}